Make recording dependencies handle directories
This allows you to use directories both for the dependency and the
files that depend on it:

  record_dependency(dir, path)

will mean that for any change to a file in dir, also reload path.
This is true even if a file is added to dir later.

  record_dependency(path, dir)

means that for any change to path, reload any files in dir.  This
is also true even if a file is added to dir later.

Note that if you are recording a dependency on a directory, you
should be requiring the directory, not the files inside it directly.