Improve logging
Before, it would log Reloading instead of Loading if the file
was already loaded, but it was kind of hard to figure out from
the logs what was happening.  Now have it log Unloading for files,
starting with the current file and any files it required. When the
unloading is finished and it is ready to load the file again, it
will log Loading.

This doesn't change the amount of logging, but I think it makes
the logs easier to understand.