Use a single Monitor instead of per-instance Mutex
Thread.exclusive was basically a single mutex, so just in case
someone is crazy and has multiple reloaders in their app in
different threads, only one will be able to reload simultaneously.

Use a Monitor instead of a Mutex, just in case someone is crazy
and tries calling the app from inside the app.  The performance
hit isn't an issue in this case.