Add Unreloader#strip_path_prefix, designed to support chrooting, bump version to 1.6.0
Previously, using rack-unreloader in chrooted environments was
kind of a pain.  This allows rack-unreloader to be used after
an application has been fully loaded, and still pick up changes
correctly to files inside the chroot.

As chroot requires root access, there is a separate spec file
for strip_path_prefix support, which forks a process per spec
to isolate changes.