Revamp README Intro & Usage
Originally, my intention was solely
to clarify a small misunderstanding I had
based on the phrasing of the README:

I thought Rack::Unreloader was supposed to watch for changes in the fs
and reload code as soon as a file was saved.
After some experimentation, however,
I learned that code is only reloaded in response to incoming HTTP requests.

As you can see, I got a little carried away with the rewrite.
In any case, I believe that this revision does a better job
of introducing the overall concept with minimal complexity up front,
discarding a code sample
that we'd never expect users to actually follow in the first place
(namely, the `require 'roda'` line in the first example).