Add support for a :delete_hook option to Unreloader#require, called when a file is deleted
This is mostly useful when loading a directory of files, where each
file updates a shared data structure such as a hash.  Adding a new
file will add an entry to the hash, and modifying and reloading the
file will update the entry in the hash, but deleting the file will
previously not update the hash, resulting in a stale entry. This
allows for a hook so that the related entry can be deleted from
the shared data structure.

To implement this, store the delete_hook for each monitored
directory and file.  This requires switching the @monitored_files
in the reloader from just storing the modify timestamp to storing
an array.