Drop Unreloader#strip_path_prefix
This is only useful in chroot situations, and the complexity isn't
worth it, considering that I doubt anyone uses it, and it is not
needed with alternative file system hiding implementations such
as unveil.  Additionally, the tests for this can only be run as
the superuser (since the tests chroot), so they weren't run
routinely, and they are now broken.