Remove guards in safe_load
safe_load is only called from reload! and require_dependencies.
reload used the FORCE options, and require_dependencies always
used a nil timestamp, so the second guard clause was unnecessary.

require_dependencies makes sure that @monitor_files has an entry
for the key.  It's likely that reload! enforces the same check,
but just to be sure, move the related guard to reload!.

Remove FORCE constant as it is now unused.